use base for vital dynamics
